feat: 메모 수정 시 이미지 미리보기 및 0.5 단위 별점 지원
- 이미지 URL 입력 → 이미지 미리보기로 변경
- 최대 2개 이미지 추가/삭제 기능 구현
- 각 이미지에 삭제 버튼 제공
- 별점 시스템 Int → Float 변경 (0.5 단위 지원)
- HalfStarRating 컴포넌트 신규 생성
- 반별(half-star) 시각적 표시 지원
- Domain/Data Layer 전체 Float 타입으로 통일

+++ frontend/app/src/main/java/com/dailymemo/data/models/request/CreateMemoRequestDto.kt
@@ -10,7 +10,7 @@ data class CreateMemoRequestDto(
     @SerializedName("image_url")
     val imageUrl: String? = null,
     @SerializedName("rating")
-    val rating: Int = 0,
+    val rating: Float = 0f,
     @SerializedName("is_pinned")
     val isPinned: Boolean = false,
     @SerializedName("latitude")
